refactor package for integration test
